requirements.txt
mytextgrid
click
PyYAML
ftfy